refactor: pack과 관련된 전역상태 이동
